Support multiple versions of same compiler vendor in site config (especially oneapi) (#1541)
This PR updates the documentation for maintaining site configurations with multiple versions of the same compiler vendor. The most important use case here is a oneapi@2024 configuration with ifort and a oneapi@2025 configuration with ifx. It turns out that this is already supported by spack-stack, therefore this is primarily a question of providing adequate documentation and updating the site configs as needed.

This PR, in addition to updating the documentation, updates the NRL site configs and NRL batch install script accordingly. These can serve as blueprints for other sites, if needed.

I also sneaked in a tiny update for the util/nrl/batch_install.sh script to turn on or off the updating of source and bootstrap caches (tested on NRL systems).