show_duplicate_packages.py: use spack.lock instead of 'spack concretize' output (#1601)
This PR updates show_duplicate_packages.py to use spack.lock to detect duplicates rather than spack concretize stdout. It is safer (especially if an environment has been concretized in multiple steps), it simplifies the code, and it's cleaner. It removes the -d option (it now only shows duplicates), and the -c option (which at this point we don't really have any use for because we're doing one-compiler environments).