Add scripts to download cargo dependencies and store in local cargo registry for offline installation (#1618)
This PR adds scripts to fetch cargo dependencies into a local registry for cached or offline installation, and for downloading and installing rust/cargo if necessary. Credits to @AlexanderRichert-NOAA for the initial version of the fetch-cargo logic.

The new logic is also incorporated in the utility utils/nrl/batch_install.sh to demonstrate how they fit into the sequence of steps.