Remove parallel-netcdf dependency except for jedi-mpas-env/unified-dev/skylab-dev (#1621)
Remove parallel-netcdf dependency from base-env. Instead, add it to jedi-mpas-env and enable the variant +pnetcdf for parallelio in jedi-mpas-env (the default for this variant is ~pnetcdf).

This means that environments using template neptune-dev concretize without parallel-netcdf, but any template that includes jedi-mpas-env (e.g. unified-dev, skylab-dev) concretize with parallel-netcdf.

I added neptune-dev as an additional template to the Ubuntu GNU runner, but had to disable the xnrl variant since this package isn't publicly available. For neptune-dev, parallel-netcdf is removed, for unified-dev and skylab-dev, it is included as before.