Add util/check_libirc.sh to check for libirc.so used in spack-stack environments (#1639)
This PR adds a utility util/check_libirc.so that checks shared libraries and executables in a spack-stack environment for links to libirc.so. The documentation adds more details on why this is needed and when/how to use it, and I added the tool to the existing GitHub actions tests. The purpose of this utility is to catch issues with spack-stack environments only.

I also created a wiki page that will provide details on how to avoid linking to libirc.so in downstream applications: https://github.com/JCSDA/spack-stack/wiki/Intel-oneAPI-compilers-and-libirc.so

The newly added check found libirc.so in several met binaries. I fixed that and improved the existing fix for parallel-netcdf in JCSDA/spack#543.