Add neptune-env to list of allowed duplicate packages in util/nrl/batch_install.sh and util/weekly_build/03_SetupEnv.sh to enable debug and release builds of it (#1682)
Add neptune-env to list of allowed duplicate packages in util/nrl/batch_install.sh and util/weekly_build/03_SetupEnv.sh to enable debug and release builds of it. This is necessary because of #1673. A similar change was made for the GitHub actions testing on the Ubuntu CI runner as part of #1673.