Add NAS as tier1 site (#1708)
This PR adds NAS as a new site for spack-stack. These files were used to build spack-stack 1.9.0. The two stacks I was able to build are oneapi@2024.2.0 with MPT and gcc@12.3.0 with Open MPI.

Note: For gcc@12.3.0 I just could not build hdf-eos2 so I had to alter the met variants to skip that.