Final updates of develop from release 1.9.3 (#1790)
Final round of updates from release 1.9.3 (branch release/1.9.0) for develop. An attempt will be made to pull these updatesinto the new feature/update_to_spack_v1 branch for changes that are not related to site configs, and for site configs that haven't been updated yet to work with spack v1.